+++ plugins/main/public/components/settings/about/appInfo.tsx
@@ -1,10 +1,24 @@
 import React from 'react';
-import { EuiCallOut, EuiFlexGroup, EuiFlexItem, EuiText } from '@elastic/eui';
+import {
+  EuiCallOut,
+  EuiFlexGroup,
+  EuiFlexItem,
+  EuiText,
+  EuiTextColor,
+  EuiTitle,
+} from '@elastic/eui';
 
-export const SettingsAboutAppInfo = ({ appInfo }: { appInfo: string }) => {
+export const SettingsAboutAppInfo = ({
+  appInfo,
+  clusterUuid,
+}: {
+  appInfo: string;
+  clusterUuid?: string | null;
+}) => {
   return (
     <EuiCallOut>
       <EuiFlexGroup
+        direction='row'
         alignItems='center'
         justifyContent='flexStart'
         gutterSize='none'
@@ -12,6 +26,18 @@ export const SettingsAboutAppInfo = ({ appInfo }: { appInfo: string }) => {
         <EuiFlexItem>
           <EuiText>
             App version: <b>{appInfo}</b>
+          </EuiText>
+        </EuiFlexItem>
+        <EuiFlexItem>
+          <EuiText>
+            Cluster UUID:{' '}
+            <b>
+              {clusterUuid ? (
+                clusterUuid
+              ) : (
+                <EuiTextColor color='subdued'>-</EuiTextColor>
+              )}
+            </b>
           </EuiText>
         </EuiFlexItem>
       </EuiFlexGroup>